Add Load, PlayPause prototypes
Work has begun and is in full swing for loading and playing (and pausing) audio files, playlists and podcasts. There is no support for keeping track of remainedTime at the moment, nor for switching to next songs/episodes in a playlist/podcast.
Loading also doesn't resume from the last place in the podcast currently.

Several dummy commands have been added to prevent crashes, though testing is limited to 01 and 02 anyway.