Add GetTop5Artists, Update code
Added GetTop5Artists to pass the statistics tests.
Updated code in various places for bugfix purposes.

Some tests do not pass, 08 for some yet-unexplained reason and 09 has a different, possibly-not-intended implementation used by the code behind the ref, where the album is expected removed even if a song from the album is contained within a loaded playlist. Not updating yet until I receive a proper response from the upstream maintainers.